# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_osi-oss_osi_internal_dto.CreatePositionRequest:
    description: Данные для создания новой должности в организации
    properties:
      department_id:
        description: |-
          ID отдела для привязки (опционально)
          Example: 1
        example: 1
        type: integer
      description:
        description: |-
          Описание должностных обязанностей
          Example: Работа с клиентами, заключение договоров
        example: Работа с клиентами, заключение договоров
        type: string
      is_admin:
        description: |-
          Является ли позиция административной (имеет расширенные права)
          Example: false
        example: false
        type: boolean
      name:
        description: |-
          Название должности (обязательное)
          Example: Менеджер по продажам
        example: Менеджер по продажам
        type: string
    required:
    - name
    type: object
...
  github_com_osi-oss_osi_internal_dto.ErrorResponse:
    description: Стандартный формат ответа при ошибке
    properties:
      error:
        description: |-
          Сообщение об ошибке
          Example: invalid email or password
        example: invalid email or password
        type: string
    type: object
...
  github_com_osi-oss_osi_internal_dto.MessageResponse:
    description: Стандартный формат успешного ответа с сообщением
    properties:
      message:
        description: |-
          Сообщение о результате операции
          Example: operation completed successfully
        example: operation completed successfully
        type: string
    type: object
...
  github_com_osi-oss_osi_internal_dto.PositionResponse:
    description: Полная информация о должности
    properties:
      created_at:
        description: |-
          Дата создания
          Example: 2024-01-15T10:30:00Z
        example: "2024-01-15T10:30:00Z"
        type: string
      department_id:
        description: |-
          ID отдела (null если не привязана к отделу)
          Example: 1
        example: 1
        type: integer
      description:
        description: |-
          Описание должности
          Example: Работа с клиентами
        example: Работа с клиентами
        type: string
      id:
        description: |-
          Уникальный идентификатор позиции
          Example: 1
        example: 1
        type: integer
      is_admin:
        description: |-
          Является ли административной
          Example: false
        example: false
        type: boolean
      name:
        description: |-
          Название должности
          Example: Менеджер по продажам
        example: Менеджер по продажам
        type: string
      organization_id:
        description: |-
          ID организации
          Example: 1
        example: 1
        type: integer
      updated_at:
        description: |-
          Дата последнего обновления
          Example: 2024-01-20T15:45:00Z
        example: "2024-01-20T15:45:00Z"
        type: string
    type: object
  github_com_osi-oss_osi_internal_dto.PositionsListResponse:
    description: Список позиций организации или отдела
    properties:
      positions:
        description: Массив позиций
        items:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionResponse'
        type: array
    type: object
...
  github_com_osi-oss_osi_internal_dto.UpdatePositionRequest:
    description: Данные для обновления должности
    properties:
      department_id:
        description: |-
          Новый ID отдела
          Example: 2
        example: 2
        type: integer
      description:
        description: |-
          Новое описание
          Example: Руководство отделом продаж
        example: Руководство отделом продаж
        type: string
      is_admin:
        description: |-
          Административная позиция
          Example: true
        example: true
        type: boolean
      name:
        description: |-
          Новое название должности
          Example: Старший менеджер по продажам
        example: Старший менеджер по продажам
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления организациями, локациями, отделами и позициями
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OSI API
  version: "1.0"
paths:
...
  /organizations/{orgId}/locations/{locId}/departments/{deptId}/positions:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает все должности, привязанные к конкретному отделу.
        Требуется доступ к организации.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID локации
        example: 1
        in: path
        minimum: 1
        name: locId
        required: true
        type: integer
      - description: ID отдела
        example: 1
        in: path
        minimum: 1
        name: deptId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список должностей отдела
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionsListResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список должностей отдела
      tags:
      - Должности
  /organizations/{orgId}/positions:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает все должности организации.
        Включает как привязанные к отделам, так и общие позиции.
        Требуется доступ к организации.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список должностей организации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionsListResponse'
        "400":
          description: Неверный формат ID организации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список всех должностей организации
      tags:
      - Должности
    post:
      consumes:
      - application/json
      description: |-
        Создаёт новую должность в организации.
        Можно привязать к отделу (department_id) или оставить общей для организации.
        Административные позиции (is_admin=true) имеют расширенные права.
        Требуется право "positions.create" или статус основателя.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: Данные для создания должности
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Должность успешно создана
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionResponse'
        "400":
          description: 'Ошибка валидации: name обязательно'
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права positions.create
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Организация или отдел не найден
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой должности
      tags:
      - Должности
  /organizations/{orgId}/positions/{posId}:
    delete:
      consumes:
      - application/json
      description: |-
        Удаляет должность из организации.
        **ВНИМАНИЕ**: Нельзя удалить должность, если на ней есть сотрудники.
        Требуется право "positions.delete" или статус основателя.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID должности
        example: 1
        in: path
        minimum: 1
        name: posId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Должность успешно удалена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.MessageResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права positions.delete
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Должность не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление должности
      tags:
      - Должности
    get:
      consumes:
      - application/json
      description: |-
        Возвращает полную информацию о должности.
        Требуется доступ к организации.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID должности
        example: 1
        in: path
        minimum: 1
        name: posId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные должности
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionResponse'
        "400":
          description: Неверный формат ID
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Должность не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение должности по ID
      tags:
      - Должности
    put:
      consumes:
      - application/json
      description: |-
        Обновляет данные должности.
        Можно изменить название, описание, привязку к отделу и административный статус.
        Требуется право "positions.update" или статус основателя.
      parameters:
      - description: ID организации
        example: 1
        in: path
        minimum: 1
        name: orgId
        required: true
        type: integer
      - description: ID должности
        example: 1
        in: path
        minimum: 1
        name: posId
        required: true
        type: integer
      - description: Новые данные должности
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.UpdatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Должность успешно обновлена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.PositionResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "401":
          description: Отсутствует или невалидный токен авторизации
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "403":
          description: Нет права positions.update
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "404":
          description: Должность не найдена
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_osi-oss_osi_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление должности
      tags:
      - Должности
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"